requirements.txt
openai
pandas
pillow
opencv-python
easyocr
pyautogui
streamlit
click
